Move filter_out_schedulable to a separate package
