+++ cluster-autoscaler/core/filteroutschedulable/filter_out_schedulable.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package core
+package filteroutschedulable
 
 import (
 	"sort"
